+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Renamed Scale.degree to Scale.getPitchByDegree
 
+- Renamed Chord.getPitch to Chord.getPitchByIndex
+
 ### Removed
 
 - Scale.getIndex